added GLLogger
GL Logger class now initialized with  Game.RunMode = RunMode.Debug
