+++ shaders/label.shaders
@@ -1,4 +1,6 @@
 [vertex]
+#version 440
+
 layout (location = 0) in vec4 vertex; // <vec2 pos, vec2 tex>
 out vec2 TexCoords;
