+++ src/lib/TauBetaChi/Interfaces/IGame.cs
@@ -1,3 +1,4 @@
+using ForgeWorks.GlowFork;
 using ForgeWorks.RailThorn;
 
 namespace ForgeWorks.TauBetaDelta.Extensibility;
@@ -6,6 +7,7 @@ public interface IGame
 {
     string Name { get; }
     string Title { get; }
+    RunMode RunMode { get; }
 
     void ChangeState(string nextState);
     IGameState GetState();
