+++ src/lib/TauBetaChi/Logging/LogLevel.cs
@@ -5,5 +5,6 @@ public enum LogLevel
     Verbose = 0,
     Debug = 1,
     Default = 2,
-    Error = 3
+    Error = 3,
+    GLDebug = 4,
 }
