+++ src/lib/TauBetaChi/RailThorn/Controls/Control.cs
@@ -1,9 +1,13 @@
 using OpenTK.Mathematics;
 
+using ForgeWorks.TauBetaDelta.Logging;
+
 namespace ForgeWorks.RailThorn.Controls;
 
 public abstract class Control
 {
+    protected static readonly LoggerManager LOGGER = LoggerManager.Instance;
+
     public string Name { get; }
     public Vector2i Location { get; set; } = (0, 0);
